+++ astminer-cli/cli.sh
