+++ astminer-cli/data/testData/arrayCalls/DoubleArrayInitialization.java
