+++ astminer-cli/data/testData/arrayCalls/IntArrayInitialization.java
