+++ astminer-cli/data/testData/arrayCalls/StringArrayInitialization.java
