+++ astminer-cli/data/testData/examples/cpp/realExamples1/2.cpp
