+++ astminer-cli/data/testData/examples/cpp/realExamples2/3.cpp
