+++ astminer-cli/data/testData/examples/cpp/realExamples2/4.cpp
