+++ astminer-cli/data/testData/examples/cpp/realExamples2/5.cpp
