+++ astminer-cli/gradle.properties
