+++ astminer-cli/src/main/kotlin/cli/Main.kt
@@ -1,4 +1,4 @@
-package examples
+package cli
 
 fun main(args: Array<String>) {
     if (args.isEmpty()) {
