+++ astminer-cli/src/main/kotlin/cli/ProjectParser.kt
@@ -1,4 +1,4 @@
-package examples
+package cli
 
 import astminer.ast.VocabularyAstStorage
 import astminer.common.Node
